+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - add `titleTextAttributes` for any `UIControl.State`.
 
+### Added
+- add SwiftUI wrapper `MultiSegmentPicker`.
+
 ### Fixed
 - fix stretched symbol font images.
 - use superview background color for selected foreground color.